core: add bootstrap world/schedule builders for shells
